+++ R/wrangle_data.R
@@ -6,7 +6,7 @@
 #' @param name_raw_json The column name in which stores user's raw data in
 #'   format of json string.
 #' @param name_raw_parsed The name used to store parsed data.
-#' @return A [tibble][tibble::tibble-package] contains the parsed data.
+#' @return A [data.frame] contains the parsed data.
 #' @export
 wrangle_data <- function(data,
                          name_raw_json = "game_data",